Refactor gqlTest into proper classes.
