+++ test/controllers/GqlTest.scala
@@ -1,9 +1,9 @@
 package controllers
 
 import akka.util.Timeout
-import controllers.GqlItTestInputs.{diseaseGenerator, geneGenerator, searchGenerator, sizeGenerator, targetDiseaseSizeGenerator}
 import controllers.GqlTest.{createRequest, generateQueryString, getQueryFromFile}
 import controllers.api.v4.graphql.GraphQLController
+import inputs.{GqlCase, GqlItTestInputs, Search, SearchPage, Target, TargetDisease, TargetDiseaseSize, TargetDiseaseSizeCursor}
 import org.scalacheck.{Gen, Shrink}
 import org.scalatestplus.play.PlaySpec
 import org.scalatestplus.play.guice.GuiceOneAppPerTest
@@ -20,123 +20,10 @@ import java.util.concurrent.TimeUnit
 import scala.concurrent.Future
 import scala.reflect.io.File
 
-object GqlItTestInputs {
 
-  lazy val geneInputs = File(this.getClass.getResource(s"/gqlInputs/genes.txt").getPath).lines.toList
-  lazy val diseaseInputs = File(this.getClass.getResource(s"/gqlInputs/efos.txt").getPath).lines.toList
-  lazy val drugInputs = Seq()
 
-  // Generators
-  val geneGenerator: Gen[String] = Gen.oneOf(geneInputs)
-  val diseaseGenerator: Gen[String] = Gen.oneOf(diseaseInputs)
-  val sizeGenerator: Gen[Int] = Gen.chooseNum(1, 10)
-  val targetDiseaseSizeGenerator: Gen[(String, String, Int)] = for {
-    gene <- geneGenerator
-    disease <- diseaseGenerator
-    size <- sizeGenerator
-  } yield (gene, disease, size)
 
-  val searchGenerator: Gen[String] = Gen.oneOf(geneGenerator, diseaseGenerator)
-}
 
-sealed trait GqlCase[T] {
-  val file: String
-  val inputGenerator: Gen[T]
-
-  def generateVariables(inputs: T): String
-}
-
-case class Search(file: String) extends GqlCase[String] {
-  override val inputGenerator = searchGenerator
-
-  def generateVariables(searchTerm: String): String = {
-    s"""
-      "variables": {
-      "queryString": "$searchTerm"
-    }
-    """
-  }
-}
-
-case class SearchPage(file: String) extends GqlCase[(String, String, Int)] {
-  val inputGenerator = for {
-    query <- searchGenerator
-    entities <- Gen.atLeastOne(Seq("target", "disease", "drug")).map(_.mkString("[\"", "\", \"", "\"]"))
-    page <- sizeGenerator
-  } yield (query, entities, page)
-
-  def generateVariables(inputs: (String, String, Int)): String =
-    s"""
-      "variables": {
-      "queryString": "${inputs._1}",
-      "index": ${inputs._3},
-      "entityNames": ${inputs._2}
-    }
-    """
-}
-
-case class Target(file: String) extends GqlCase[String] {
-
-  val inputGenerator = geneGenerator
-
-  def generateVariables(target: String): String = {
-    s"""
-      "variables": {
-      "ensgId": "$target"
-    }
-    """
-  }
-}
-
-case class TargetDisease(file: String) extends GqlCase[(String, String)] {
-  val inputGenerator = for {
-    gene <- geneGenerator
-    disease <- diseaseGenerator
-  } yield (gene, disease)
-
-  def generateVariables(inputs: (String, String)): String = {
-    val (target, disease) = inputs
-    s"""
-      "variables": {
-      "efoId": "$target",
-      "ensemblId": "$disease"
-    }
-    """
-  }
-}
-
-case class TargetDiseaseSize(file: String) extends GqlCase[(String, String, Int)] {
-  val inputGenerator = targetDiseaseSizeGenerator
-
-  def generateVariables(input: (String, String, Int)): String = generateVariables(input._1, input._2, input._3)
-
-  def generateVariables(target: String, disease: String, size: Int): String = {
-    s"""
-      "variables": {
-      "efoId": "$target",
-      "ensemblId": "$disease",
-      "size": $size
-    }
-    """
-  }
-}
-
-case class TargetDiseaseSizeCursor(file: String) extends GqlCase[(String, String, Int)] {
-  val inputGenerator = targetDiseaseSizeGenerator
-
-  def generateVariables(input: (String, String, Int)): String = generateVariables(input._1, input._2, input._3)
-
-  def generateVariables(target: String, disease: String, size: Int): String = {
-    s"""
-    "variables": {
-      "efoId": "$target",
-      "ensemblId": "$disease",
-      "size": $size,
-      "cursor": null
-    }
-  """
-  }
-}
 
 object GqlTest {
